EmployeeManagementSystem/ems-chatbot/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
groq
python-dotenv

pandas
plotly
sqlalchemy
pymysql
psycopg2-binary
requests